fix: nip and password data type

+++ internal/domain/user-it.go
@@ -9,14 +9,14 @@ import (
 type UserIT struct {
 	ID        string    `db:"id"`
 	CreatedAt time.Time `db:"created_at"`
-	NIP       int       `db:"nip"`
+	NIP       int64     `db:"nip"`
 	Name      string    `db:"name"`
 	Password  string    `db:"password"`
 	Role      string    `db:"role"`
 }
 
 type RegisterUserIT struct {
-	NIP      *int   `json:"nip" validate:"required,nip=it"`
+	NIP      *int64 `json:"nip" validate:"required,nip=it"`
 	Name     string `json:"name" validate:"required,min=5,max=50"`
 	Password string `json:"password" validate:"required,min=5,max=33"`
 }
@@ -28,7 +28,7 @@ type LoginUserIT struct {
 
 type GetUserIT struct {
 	ID        string    `json:"userId"`
-	NIP       int       `json:"nip"`
+	NIP       int64     `json:"nip"`
 	Name      string    `json:"name"`
 	CreatedAt time.Time `json:"createdAt"`
 }
